Aggiunta logica per gestione prezzo gerarchico con MediatR

+++ GestHotelsDomain/Repositories/Price/PriceRepository.cs
@@ -1,5 +1,6 @@
 ﻿using GestHotelsDomain.Data;
 using GestHotelsDomain.Entities;
+using GestHotelsDomain.Services;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -18,24 +19,41 @@ namespace GestHotelsDomain.Repositories.Price
             _context = dbContext;
         }
 
-        public async Task<Entities.Price> AddPriceAsync(Entities.Price price)
+        public async Task<string> AddPriceAsync(Entities.Price price)
         {
+            if (_context.Price == null)
+            {
+                return "Entity set 'HotelDbContext.Price' is null.";
+            }
             var result = _context.Price.Add(price);
-            await _context.SaveChangesAsync();
-            return result.Entity;
+            HierarchyHelper hierarchyHelper = new HierarchyHelper(_context, price);
+            if (hierarchyHelper.ValidatePrice())
+            {
+                await _context.SaveChangesAsync();
+                return String.Format("New Price with id = {0} created succesfully",result.Entity.Id);
+            }
+            else
+            {
+                return "The price does not respect the hierarchy rules";
+            }
         }
 
-        public async Task<int> DeletePriceAsync(int Id)
+        public async Task<string> DeletePriceAsync(int Id)
         {
             var filteredData = _context.Price.Where(x => x.Id == Id).FirstOrDefault();
+            if (filteredData == null)
+            {
+                return String.Format("Price with id = {0} not found", Id);
+
+            }
             _context.Price.Remove(filteredData);
-            return await _context.SaveChangesAsync();
+             await _context.SaveChangesAsync();
+            return string.Format("Price with id = {0} deleted succesfully", Id);
         }
 
         public async Task<Entities.Price> GetPriceByIdAsync(int id)
         {
-            return await _context.Price
-                            .FirstOrDefaultAsync(h => h.Id == id);
+            return await _context.Price.FirstOrDefaultAsync(h => h.Id == id);
         }
 
         public async Task<List<Entities.Price>> GetPriceListAsync()
@@ -44,10 +62,54 @@ namespace GestHotelsDomain.Repositories.Price
                             .ToListAsync();
         }
 
-        public async Task<int> UpdatePriceAsync(Entities.Price price)
+        public async Task<string> UpdatePriceAsync(Entities.Price price)
         {
+            //var price = await _context.Price.FindAsync(id);
+            if (price == null)
+            {
+                return String.Format("Price with id = {0} not found",price.Id);
+            }
             _context.Price.Update(price);
-            return await _context.SaveChangesAsync();
+            HierarchyHelper hierarchyHelper = new HierarchyHelper(_context, price);
+            if (hierarchyHelper.ValidatePrice())
+            {
+                await _context.SaveChangesAsync();
+                return String.Format("Price with id = {0} updated succesfully", price.Id);
+
+            }
+            else
+            {
+                return "The price does not respect the hierarchy rules";
+            }
+            //price.Modify(newPrice);
+            //_context.Entry(price).State = EntityState.Modified;
+
+            //try
+            //{
+            //    HierarchyHelper hierarchyHelper = new HierarchyHelper(_context, price);
+            //    if (hierarchyHelper.ValidatePrice())
+            //    {
+            //        await _context.SaveChangesAsync();
+            //        return CreatedAtAction("GetPrice", new { id = price.Id }, price);
+            //    }
+            //    else
+            //    {
+            //        return Problem("The price does not respect the hierarchy rules");
+            //    }
+            //}
+            //catch (DbUpdateConcurrencyException)
+            //{
+            //    if (!PriceExists(id))
+            //    {
+            //        return NotFound();
+            //    }
+            //    else
+            //    {
+            //        throw;
+            //    }
+            //}
+            //_context.Price.Update(price);
+            //return await _context.SaveChangesAsync();
         }
     }
 }
